pruebatecno/requirements.txt
Django == 5.2.10
gunicorn
psycopg2-binary
django-prometheus
psutil
influxdb-client
PyYAML>=6.0
whitenoise>=6.0